Update PalindromeChecker.java
This is a simple Java program that checks to see if a number entered is a palindrome. Whatever is input is stored as the string object "userNum". A StringBuilder object is created and initialized to the contents of the string userNum. The StringBuilder method reverse() is used to reverse the characters of the StringBuilder object before the toString() method is used to return a string with the data. Finally, an if statement is used to determine whether the reverse string is equal to the one input by the user. If so, the output specifies. If not, the argument for the else statement is satisfied and the string is not a palindrome. The output also specifies this to the user. Technically, this works for more than just numbers. Since a string is being entered, any sequence of characters such as a word or a name will still work.