Add files via upload
This is another version of the Race Selection program that attempts to use method calls for each of the races and places the information for each in a different class. The final project will combine the race selection, stat builder, and some sort of challenge for the character to overcome.